feat(multiplayers) add second player locally; add routes to access home and game
