+++ src/game/index.js
@@ -1,0 +1,1 @@
+export { Game } from "./Game";
